progress towards Completing Temperature Conversion
some outstanding issues, first and foremost, getTemperature range needs to be able to discern between 1, 2, 3,4 and so does display data.

After that it's nothing other than organization.

Also, I should consult Mr. Marshall about this code because this is the most "spaghettii" like code I've probably ever written (probably not but it feels like it)